requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
construct==2.10.68
future==0.18.3
idna==3.4
keepass==1.2
lxml==4.9.2
pycparser==2.21
pycrypto==2.6.1
pycryptodomex==3.17
pykeepass==4.0.3
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.2
six==1.16.0
urllib3==1.26.15
